+++ packages/driver/cypress/e2e/dom/visibility.cy.ts
@@ -169,6 +169,10 @@ describe('src/cypress/dom/visibility', () => {
     })
   })
 
+  const add = (el) => {
+    return $(el).appendTo(cy.$$('body'))
+  }
+
   context('hidden/visible overrides', () => {
     beforeEach(function () {
       // ensure all tests run against a scrollable window
@@ -178,8 +182,8 @@ describe('src/cypress/dom/visibility', () => {
       this.$parentVisHidden = add('<div class="invis" style="visibility: hidden;"><button>parent visibility: hidden</button></div>')
       this.$displayNone = add('<button style="display: none">display: none</button>')
       this.$inputHidden = add('<input type="hidden" value="abcdef">')
-      this.$divNoWidth = add('<div style="width: 0; height: 100px;">width: 0</div>')
-      this.$divNoHeight = add('<div style="width: 50px; height: 0px;">height: 0</div>')
+      this.$divNoWidth = add('<div style="width: 0; height: 100px;"></div>')
+      this.$divNoHeight = add('<div style="width: 50px; height: 0px;"></div>')
       this.$divDetached = $('<div>foo</div>')
       this.$divVisible = add(`<div>visible</div>`)
 
@@ -691,7 +695,6 @@ describe('src/cypress/dom/visibility', () => {
         cy.wrap(this.$optionHiddenInSelect.find('#hidden-opt')).should('not.be.visible')
       })
 
-      // TODO(webkit): fix+unskip
       it('follows regular visibility logic if option outside of select', { browser: '!webkit' }, function () {
         expect(this.$optionOutsideSelect.find('#option-hidden').is(':hidden')).to.be.true
         expect(this.$optionOutsideSelect.find('#option-hidden')).to.be.hidden
@@ -740,7 +743,57 @@ describe('src/cypress/dom/visibility', () => {
     })
 
     describe('width and height', () => {
-      it('is hidden if offsetWidth is 0', function () {
+      it('is visible when el.textContent, even if offsetWidth is 0', function () {
+        this.$divNoWidthTextContent = add('<div style="width: 0; height: 100px;">width: 0</div>')
+
+        expect(this.$divNoWidthTextContent.is(':hidden')).to.be.false
+        expect(this.$divNoWidthTextContent.is(':visible')).to.be.true
+
+        expect(this.$divNoWidthTextContent).to.not.be.hidden
+        expect(this.$divNoWidthTextContent).to.be.visible
+
+        cy.wrap(this.$divNoWidthTextContent).should('be.not.hidden')
+        cy.wrap(this.$divNoWidthTextContent).should('be.visible')
+      })
+
+      it('is visible when el.textContent, even if offsetHeight is 0', function () {
+        this.$divNoHeightTextContent = add('<div style="width: 50px; height: 0px;">height: 0</div>')
+
+        expect(this.$divNoHeightTextContent.is(':hidden')).to.be.false
+        expect(this.$divNoHeightTextContent.is(':visible')).to.be.true
+
+        expect(this.$divNoHeightTextContent).to.not.be.hidden
+        expect(this.$divNoHeightTextContent).to.be.visible
+
+        cy.wrap(this.$divNoHeightTextContent).should('be.not.hidden')
+        cy.wrap(this.$divNoHeightTextContent).should('be.visible')
+      })
+
+      it('is hidden when when el.textContent contains only whitespace and offsetWidth is 0', function () {
+        this.$divNoHeightBlankTextContent = add('<div style="width: 0px; height: 50px;">   \n   \t  </div>')
+
+        expect(this.$divNoHeightBlankTextContent.is(':hidden')).to.be.true
+        expect(this.$divNoHeightBlankTextContent.is(':visible')).to.be.false
+
+        expect(this.$divNoHeightBlankTextContent).to.be.hidden
+        expect(this.$divNoHeightBlankTextContent).to.not.be.visible
+
+        cy.wrap(this.$divNoHeightBlankTextContent).should('be.hidden')
+        cy.wrap(this.$divNoHeightBlankTextContent).should('not.be.visible')
+      })
+
+      it('is hidden when no el.textContent with offsetHeight is 0', function () {
+        expect(this.$divNoHeight.is(':hidden')).to.be.true
+        expect(this.$divNoHeight.is(':visible')).to.be.false
+
+        expect(this.$divNoHeight).to.be.hidden
+        expect(this.$divNoHeight).to.not.be.visible
+
+        cy.wrap(this.$divNoHeight).should('be.hidden')
+        cy.wrap(this.$divNoHeight).should('not.be.visible')
+      })
+
+      it('is hidden when no el.textContent with offsetWidth is 0', function () {
         expect(this.$divNoWidth.is(':hidden')).to.be.true
         expect(this.$divNoWidth.is(':visible')).to.be.false